Initial commit for adding case for article deletion.

+++ app/controllers/articles_controller.rb
@@ -48,9 +48,16 @@ class ArticlesController < ApplicationController
   end
   
   def destroy
-    if @article.destroy
-      flash[:success] = "Article has been deleted"
-      redirect_to articles_path
+    unless @article.user == current_user
+      flash[:alert] = "You can only delete your own article."
+      redirect_to root_path
+    else
+      if @article.destroy
+        flash[:success] = "Article has been deleted."
+        redirect_to articles_path
+      else
+        flash[:danger] = "There was a problem deleting the article."
+      end
     end
   end
